requirements.txt
cffi==1.15.1
click==8.1.3
cryptography==38.0.1
Flask==2.2.2
Flask-Cors==3.0.10
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql==0.0.3
mysqlclient==2.1.1
numpy==1.23.4
pandas==1.5.1
pycparser==2.21
PyMySQL==1.0.2
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.5
six==1.16.0
Werkzeug==2.2.2
zipp==3.10.0
requests==2.28.1
